% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_datafindr_files}
\alias{copy_datafindr_files}
\title{Copy templates for DataFindR}
\usage{
copy_datafindr_files(path, overwrite = FALSE)
}
\arguments{
\item{path}{Character vector. Target path}

\item{overwrite}{Boolean. Default \code{FALSE}}
}
\value{
No return value, but copies files to the specified path
}
\description{
Imports templates for prompts and schemas
}
